+++ DeltGeometriFelleskomponent.TopologyImplementation/PolygonCreator.cs
@@ -1,5 +1,6 @@
 ﻿using DeltGeometriFelleskomponent.Models;
 using NetTopologySuite.Geometries;
+using NetTopologySuite.Operation.Polygonize;
 
 namespace DeltGeometriFelleskomponent.TopologyImplementation;
 
@@ -14,6 +15,121 @@ public class PolygonCreator
         {
             AffectedFeatures = request.AffectedFeatures.Concat(new List<NgisFeature>(){request.Feature, exteriorLine}).Concat(interiorLines).ToList(),
             IsValid = true
+        };
+    }
+
+    public TopologyResponse CreatePolygonFromLines(string type, List<NgisFeature> lineFeatures, Point? centroid)
+    {
+        // use  Polygonizer polygonizer for all operations
+        var isValid = true;
+        Polygon? polygon = null;
+        // Now supports  multiple linestrings and order
+        Polygonizer polygonizer = new Polygonizer(extractOnlyPolygonal: true);
+
+        foreach (var lineFeature in lineFeatures)
+        {
+            polygonizer.Add(lineFeature.Geometry);
+        }
+
+        if (polygonizer.GetPolygons().Count > 0)
+        {
+            polygon = (Polygon)polygonizer.GetPolygons().First();
+            isValid = polygon.IsValid;
+
+            if (!polygon.Shell.IsCCW)
+            {
+                // TODO: Check if polygon.Reverse is enough
+                Console.WriteLine("Polygon is not CCW");
+                polygon = (Polygon)polygon.Reverse(); // reverse polygon
+            }
+
+            if (centroid != null)
+            {
+                var inside = polygon.Contains(centroid);
+                Console.WriteLine("Point is inside polygon:{0}", inside);
+                isValid = inside;
+            }
+        }
+        else
+        {
+            isValid = false;
+        }
+
+        var cutEdges = polygonizer.GetCutEdges();
+        var dangels = polygonizer.GetDangles();
+
+        if (cutEdges.Count > 0) Console.WriteLine("cutEdges.Count:{0}", cutEdges.Count);
+        if (dangels.Count > 0) Console.WriteLine("dangels.Count:{0}", dangels.Count);
+
+
+        var lokalId = Guid.NewGuid().ToString();
+
+
+        var polygonFeature = NgisFeatureHelper.CreateFeature(polygon, lokalId);
+
+        if (polygon != null)
+        {
+
+
+            var exterior = polygon.ExteriorRing;
+            var interiors = polygon.InteriorRings;
+            var referencesExterior = new List<NgisFeature>();
+            List<List<NgisFeature>>? referencesInteriors = null;
+
+            foreach (var feature in lineFeatures)
+            {
+                if (feature.Geometry.CoveredBy(exterior))
+                {
+                    referencesExterior.Add(feature);
+                }
+            }
+
+            if (interiors != null && interiors.Length > 0)
+            {
+                foreach (var hole in interiors)
+                {
+                    List<NgisFeature>? referencesOnehole = null;
+                    foreach (var feature in lineFeatures)
+                    {
+                        if (feature.Geometry.CoveredBy(hole))
+                        {
+                            //referencesInteriors ??= new List<List<NgisFeature>>();
+                            if (referencesInteriors == null)
+                            {
+                                referencesInteriors = new List<List<NgisFeature>>();
+                            }
+
+                            referencesOnehole ??= new List<NgisFeature>();
+                            //if (referencesOnehole == null)
+                            //{
+                            //    referencesOnehole = new List<NgisFeature>();
+                            //}
+
+
+                            referencesOnehole.Add(feature);
+                        }
+                    }
+
+                    if (referencesOnehole != null) referencesInteriors?.Add(referencesOnehole);
+                }
+            }
+
+            NgisFeatureHelper.SetReferences(polygonFeature, referencesExterior, referencesInteriors);
+        }
+
+        else
+        {
+            NgisFeatureHelper.SetReferences(polygonFeature, lineFeatures, null);
+        }
+        NgisFeatureHelper.SetOperation(polygonFeature, Operation.Create);
+
+        lineFeatures.ForEach(a => NgisFeatureHelper.SetReferences(a, new List<string>() { lokalId }, null));
+
+        //set type to type
+        return new TopologyResponse()
+        {
+            AffectedFeatures = new List<NgisFeature>(){ polygonFeature },
+            IsValid = isValid
         };
     }
